pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
starlette = "^0.38.2"
uvicorn = "^0.30.6"
google-auth = "^2.3.3"
google-api-python-client = "^2.31.0"
jinja2 = "^3.1.4"
pydantic-settings = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
